### 2023-04-27 #5416 @editorialbot
[PRE REVIEW]: SyntheticEddyMethod.jl: A Julia package for the creation of inlet flow conditions for LES
**Submitting author:** @carlodev (<a href="http://orcid.org/0000-0002-2873-6293">Carlo Brunelli</a>)
**Repository:** https://github.com/carlodev/SyntheticEddyMethod.jl
**Branch with paper.md** (empty if default branch): JossSub
**Version:** v0.4.1
