+++ paper.bib
@@ -7,7 +7,7 @@
   pages={65--98},
   year= 2017,
   publisher={SIAM},
-  url={https://doi.org/10.1137/141000671}
+  doi={10.1137/141000671}
 }
 
 @article{Jarrin:2006,
@@ -46,7 +46,7 @@ volume = {134},
 pages = {1261-1267},
 year = {2019},
 issn = {0017-9310},
-doi = {https://doi.org/10.1016/j.ijheatmasstransfer.2019.02.061},
+doi = {10.1016/j.ijheatmasstransfer.2019.02.061},
 url = {https://www.sciencedirect.com/science/article/pii/S0017931018332289},
 author = {Geunwoo Oh and Kyung Min Noh and Hyunwook Park and Jung-Il Choi},
 keywords = {Synthetic eddy method, Inflow boundary condition, Turbulent thermal boundary layer, Large-eddy simulation},
