### 2023-05-02 206a7d4
CompatHelper: bump compat for StatsBase to 0.34, (keep existing compat)
