+++ Project.toml
@@ -33,17 +33,17 @@ FFTW = "1.5.0"
 FileIO = "1.16.0"
 Interpolations = "0.14.7"
 JLD2 = "0.4.30"
+LaTeXStrings = "1.3.0"
 MCMCChains = "5.7.1"
 Plotly = "0.4.1"
 PlotlyJS = "0.18.10"
 Plots = "1.38.5"
 PyPlot = "2.11.0"
 Revise = "3.5.1"
-StatsBase = "0.33.21"
+StatsBase = "0.33.21, 0.34"
 StatsPlots = "0.15.4"
 Trapz = "2.0.3"
 XLSX = "0.7.9, 0.8"
-LaTeXStrings = "1.3.0"
 julia = "1.8.2"
 
 [extras]
