+++ previews/PR17/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"advanced_usage/#Advanced-usage","page":"Advanced Usage","title":"Advanced usage","text":"","category":"section"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"In this page are provided some example on how the user can customize the eddy generation.","category":"page"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"Defining the Virtual Box in y and z directions.","category":"page"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"using SyntheticEddyMethod\n\nσ = 0.1\nb = 2.0\na = 0.0\n\ny = collect(a:0.1:b)\nz = collect(a:0.1:b)","category":"page"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"It is possible also to specify the x direction dimensions, a different shape funcion and different σ in the three different directions. The shape coded can be seen in the Shapefunctions.jl file.","category":"page"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"x = collect(-1:0.1:1)\nσx = 0.1\nσy = 0.05\nσz = 0.07\nσ = [σx,σy,σz]\nVboxinfo = VirtualBox(x,y,z,σ; shape_fun = step_fun)\nN = Vboxinfo.N\nEddies = initialize_eddies(Vboxinfo)\n","category":"page"},{"location":"advanced_usage/#Define-custom-Reynolds-Stress-Tensor","page":"Advanced Usage","title":"Define custom Reynolds Stress Tensor","text":"","category":"section"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"The user can also use a custom Reynolds stress just by writing a 3x3 matrix.","category":"page"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"Re = [1.87e-5 -8.6e-8  -8.2e-7; -8.6e-8 5.27e-8 6.8e-8; -8.2e-7 4.9e-9 2.64e-6]","category":"page"},{"location":"advanced_usage/#Import-the-Reynolds-Stress-Tensor-from-file","page":"Advanced Usage","title":"Import the Reynolds Stress Tensor from file","text":"","category":"section"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"Or use a database where the the Reynolds Stress is defined pointwise. ","category":"page"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"using XLSX\nreynolds_stress_file = joinpath(@__DIR__,\"..\",\"..\",\"test\",\"Data\",\"Re_ch.xlsx\")\nA_from_file = get_reynolds_stress_from_file(reynolds_stress_file)","category":"page"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"An curious user notice that in this last case the Reynolds Stress is not a matrix, but it is an interpolator object. It depends on the point location where the user want to compute the fluctuations.","category":"page"},{"location":"advanced_usage/","page":"Advanced Usage","title":"Advanced Usage","text":"vector_points =
[…]
