### 2023-06-08 #5416 @philipcardiff
> Hi Philip, Sure, I would be happy to review. -- Paul
> […](#)

Great, thank you, Paul.

Here are the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html).

I will find at least one other reviewer.
