### 2023-06-16 #5565 @atzberg
Overall, the package is well-organized with good discussions of the methods, implementations, and usage.  The package provides implementations of simulation methods from the literature for modeling turbulent flows.   Below are a few relatively minor items and typos to fix.

In the SyntheticEddyMethod.jl/README.md:

* Small typo to fix
"Then, eddies are initialize in the virtualbox... "
should be
"Then, eddies are initialized in the virtualbox... "

* A little confused by these two sentences
"Compute the velocity fluctuation. It is then is 'corrected' using the matrix A which is internally I created using the..."
if I understood correctly, I would suggest re-phrasing similar to
"For computing the velocity fluctuations, there is a 'correction' using the matrix A which is internally created using the..."
(or re-phrase to clarify what was meant by 'corrected').

In the paper.md -> PDF:

* small typo / clarify
"...and to control the level of turbulence and eddies length of the generated fields."
suggest changing
"...and to control the level of turbulence and length-scales of the eddies in the generated fields."
(or something similar to clarify 'eddies length')

* "...implementing the divergence-free (DFSEM) constraint at the fluctuations for incompressible flows, as most of the case of turbulence."
suggest changing to something similar to
"...implementing the divergence-free (DFSEM) constraint to obtain fluctuations for incompressible flows, which are the most common cases studied for turbulence."

* "Julia is extremely expressive and allows one to condensate complex mathematical expression in a few synthetic lines."
suggest editing to
"Julia is extremely expressive and allows one to condensate complex mathematical expression into a few syntactic lines."

In summary, the package is put together well with good discussions of the implemented methods and on the usage of the APIs.  After making the minor edits mentioned above, I recommend accepting for JOSS.
