### 2023-07-17 #5565 @akshaysridhar
`SyntheticEddyMethod.jl` is a Julia language implementation of methods published in literature that enable easy generation of turbulent inflow profiles. The package is generally well organised, and satisfies the key requirements listed in the review checklist. The accompanying paper is concise, and the listed features are consistent with the code. A few suggestions for improvement are listed below.

**Tests**
- `]test` out of the box run, with 84/84 passing tests in a generic group `SyntheticEddyMethod.jl` - These tests could be grouped into more useful subsections to allow future contributors to easily identify key components.
-`[ Info: convecting eddies` is a repetitive printed statement when launching `test` - is this necessary? Can this be condensed into a simpler display statement?
- Please include a statement explaining why this block is commented: https://github.com/carlodev/SyntheticEddyMethod.jl/blob/eca3abb9a4744add104a93cc6297480eed070415/test/read_reynoldsstress.jl#L32-L78 - If this indeed verifies consistency with a "known/provided" profile - perhaps this is better included within a `testset` block (or as an example).

**Documentation**
- `References` section in documentation currently contains API information. I suggest renaming this as such.
- I suggest using https://docs.juliahub.com/DocumenterCitations/B0owD/0.2.2/#Telling-Documenter.jl-about-your-bibliography to document citations using `at-cite` macro in Julia.
- Small type-error fixes: "referers -> refers", "fluctutation -> fluctuation ", "coeherent -> coherent", "speicific -> specific"

**Paper**
- Overall, the paper is concise and supports the main code documentation in its descriptions of contributions.
- The text in the `results` section could be condensed into the figure caption. The block that currently describes the figures could perhaps then be used to highlight advanced use cases.

I believe this submission is suitable for publication following the minor changes listed above!
