### 2023-07-18 b948ebe
condensation of results section of the paper
