### 2023-07-18 ef33903
add DocumenterCitations style references
