+++ docs/make.jl
@@ -1,6 +1,10 @@
-using Documenter, SyntheticEddyMethod
+using Documenter, DocumenterCitations, SyntheticEddyMethod
 
-makedocs(
+
+bib = CitationBibliography(joinpath(@__DIR__, "src", "docs.bib"); style=:numeric)
+
+
+makedocs(bib;
     sitename = "SyntheticEddyMethod.jl",
     modules = [SyntheticEddyMethod],
     pages = [
@@ -8,7 +12,8 @@ makedocs(
         "Usage" => "usage.md",
         "Advanced Usage" => "advanced_usage.md",
         "Exploring" => "exploring.md",
-        "API information" => "api_info.md"
+        "API information" => "api_info.md",
+        "References" => "references.md",
     ],
 )
 
