+++ Project.toml
@@ -6,6 +6,7 @@ version = "0.4.1"
 [deps]
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
@@ -29,10 +30,12 @@ XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
 [compat]
 DataFrames = "1.5.0"
 Documenter = "0.27.23"
+DocumenterCitations = "1.0.0"
 FFTW = "1.5.0"
 FileIO = "1.16.0"
 Interpolations = "0.14.7"
 JLD2 = "0.4.30"
+LaTeXStrings = "1.3.0"
 MCMCChains = "5.7.1"
 Plotly = "0.4.1"
 PlotlyJS = "0.18.10"
@@ -43,7 +46,6 @@ StatsBase = "0.33.21"
 StatsPlots = "0.15.4"
 Trapz = "2.0.3"
 XLSX = "0.7.9, 0.8"
-LaTeXStrings = "1.3.0"
 julia = "1.8.2"
 
 [extras]
