+++ test/TestDrivers/driver_fluctuations.jl
@@ -1,7 +1,10 @@
 
 using Statistics
 using LinearAlgebra
-function driver_test(TI::Float64; Nt = 1000, s_fun = tent_fun, DFSEM = false)
+
+
+
+function fluctuations_test(TI::Float64; Nt = 1000, s_fun = tent_fun, DFSEM = false)
 σ = 0.1 #eddy dimensions, the same in all the directions
 b = 5.0
 a = 0.0
@@ -13,20 +16,6 @@ z = collect(a:0.1:b)
 
 
 Vboxinfo = VirtualBox(y,z,σ; shape_fun = s_fun)
-
-@test Vboxinfo.σ == [σ,σ,σ]
-@test Vboxinfo.X_end == σ
-@test Vboxinfo.X_start == -σ
-@test Vboxinfo.Y_end == b+σ
-@test Vboxinfo.Y_start == a-σ
-@test Vboxinfo.Z_end == b+σ
-@test Vboxinfo.Z_start == a-σ
-
-Vboxinfo.N = 100
-@test Vboxinfo.N == 100
-
-Vboxinfo = VirtualBox(y,z,σ; shape_fun = s_fun)
-N = Vboxinfo.N
 
 dt = 0.01
 
@@ -48,11 +37,6 @@ Eddies = initialize_eddies(Vboxinfo)
 Re_stress_2, Eddies = initialize_eddies(U₀, TI, Vboxinfo)
 
 @test Re_stress == Re_stress_2
-
-#Computing the velocity in the middle of the VirtualBox domain
-vector_points = create_vector_points(x, y, z)
-
-@test length(vector_points) == length(x)*length(y)*length(z)
 
 
 #Computing the velocity in the middle of the VirtualBox domain
@@ -77,7 +61,11 @@ end
     s2 = Statistics.std(U[:,2])
     s3 = Statistics.std(U[:,3])
 
-return mean([s1,s2,s3])
+TI_computed = mean([s1,s2,s3])
+
+#TEST TI computed == TI provided by the user
+@test isapprox(TI_computed, TI; rtol =0.2)
+ 
 
 end
 
@@ -117,6 +105,3 @@ function test_div_null(tol=0.1)
 end
 
 
-function utilities_test()
-    return isapprox(compute_Ek([1.2, 0.1, 0.3], 1.0), (0.2^2 + 0.1^2 + 0.3^2 )*0.5)
-end
