+++ test/runtests.jl
@@ -2,23 +2,24 @@ using SyntheticEddyMethod
 using Test
 using Plots
 
-@testset "SyntheticEddyMethod.jl" begin
 
-    include("driver.jl")
-    @test isapprox(driver_test(0.1), 0.1; rtol =0.2)
-    @test isapprox(driver_test(0.01), 0.01; rtol =0.2)
-    @test isapprox(driver_test(0.05), 0.05; rtol =0.2)
 
-    @test isapprox(driver_test(0.01; s_fun = step_fun), 0.01; rtol =0.2)
-    @test isapprox(driver_test(0.01; s_fun = trunc_gauss_fun), 0.01; rtol =0.2)
+@testset "VirtualBox" begin include("TestDrivers/driver_virtualbox.jl") end
 
-    @test isapprox(driver_test(0.01; s_fun = DFSEM_fun, DFSEM = true), 0.01; rtol =0.2)
-
+@testset "Fluctuations" begin include("TestDrivers/driver_fluctuations.jl") 
+    fluctuations_test(0.1)
+    fluctuations_test(0.01)
+    fluctuations_test(0.05)
+    fluctuations_test(0.01; s_fun = step_fun)
+    fluctuations_test(0.01; s_fun = trunc_gauss_fun)
+    fluctuations_test(0.01; s_fun = DFSEM_fun, DFSEM = true)
     test_div_null(0.001)
+end
 
-    @test utilities_test()
-
-    include("read_reynoldsstress.jl")
+@testset "ReynoldsStress" begin include("TestDrivers/driver_reynoldsstress.jl") 
     read_re_test()
+    test_anystoropic_reynolds()
+end 
 
-end    
+
+@testset "Utilities" begin include("TestDrivers/driver_utilities.jl") end

[…]
