### 2023-07-18 #5565 @carlodev
Thank you @akshaysridhar for your insightful comments and suggestions. I have applied all the modifications, feel free to check and verify if I understood your remarks correctly. I also reorganised the `test` section, and added the commented example.

Concerning the `Info: connecting eddies` is an indication to the user that Eddies are convected in the VirtualBox. During the testing phase it is  printed continuously because the validation/test process is statistics, so a significant number of samples is needed. The statistical validation can take up quite some time, but seeing that statement printed provides feedback about the fact that is running.
