### 2023-07-18 #5565 @akshaysridhar
Thanks @carlodev for making the suggested changes. I've checked out the `JossSub` branch, and have the following suggestions:

(1) The `doc-build` with these results in the following warnings which I believe can be easily fixed by including these functions in the `at-docs` block in the api info markdown file.
```
[ Info: CheckDocument: running document checks.
┌ Warning: 2 docstrings not included in the manual:
│
│     SyntheticEddyMethod.Reynolds_stress_point :: Tuple{Vector{Float64}, Reynolds_stress_interpolator}
│     SyntheticEddyMethod.Reynolds_stress_points :: Tuple{Vector{Float64}, Reynolds_stress_interpolator}
│
│ These are docstrings in the checked modules (configured with the modules keyword)
│ that are not included in @docs or @autodocs blocks.
```
(2) "coeherent -> coherent" in `./docs/src/index.md:15:- Create coeherent eddies in 3D domain`

(3) "fluctutation -> fluctuation " in ` ./docs/src/usage.md:84:Now, we are going to evaluate the fluctutations`

(4) In
```
 21 @testset "ReynoldsStress" begin include("TestDrivers/driver_reynoldsstress.jl")
 22     read_re_test()
 23     test_anystoropic_reynolds()
 24 end
```
-> Should this be `anisotropic` (and therefore also updated in the corresponding function definition?

(5) Re `[ Info: convecting eddies` : I see that in a standard use case this is a single info statement in the `compute_fluct` function -> I think a combination of `@info <statement>  maxlog=<integer>` with an indicator of the `total number of samples collected for the test` would be a better option (This would be succinct, and retains the information about the action the code is performing in your above comment).
