+++ README.md
@@ -99,7 +99,7 @@ It is reported the normalized divergence in a plane using the DFSEM.
 ## Package Features
 - Create velocity fluctuations for inlet boundary conditions
 - Create fluctuations that respect the divergence-free condition (DFSEM)
-- Create coeherent eddies in a 3D domain
+- Create coherenteddies in a 3D domain
 - It can simulate anisotropic effects by allowing the eddies to have different dimensions along different directions
 - Define a custom Reynolds Stress Tensor
 - Import custom Reynolds Stress Tensor
