### 2023-07-19 #5565 @carlodev
Thanks @akshaysridhar for your suggestions. I have fixed the docs warnings, the typos. I also modified the `@info` macro as you suggest in order to avoid seeing `[ Info: convecting eddies` so many times
