### 2023-07-19 #5565 @philipcardiff
HI @carlodev , I have proposed a few minor English edits in a pull request: https://github.com/carlodev/SyntheticEddyMethod.jl/pull/20.

Let me know when these have been incorporated.
