### 2023-07-19 #5565 @akshaysridhar
@philipcardiff @carlodev
Apologies for missing this earlier - but the pdf file doesn't appear to render figure captions correctly - I see that these captions for the two figures are included in the code as :
```
![Spectra examples at different Turbulent Intensities using tent function as shape function. It refers to the fluctuations in time in one specific point.](images/docs/Spectra.png){ width=50% }
![normalized divergence in a plane using the divergence-free feature. \label{fig:dfsem-plane}](images/docs/Div_free_plane.png){ width=50%  }
```
