+++ dev/docs.bib
@@ -1,0 +1,53 @@
+@article{Bezanson:2017,
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal={SIAM review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year= 2017,
+  publisher={SIAM},
+  doi={10.1137/141000671}
+}
+
+@article{Jarrin2006,
+   abstract = {The generation of inflow data for spatially developing turbulent flows is one of the challenges that must be addressed prior to the application of LES to industrial flows and complex geometries. A new method of generation of synthetic turbulence, suitable for complex geometries and unstructured meshes, is presented herein. The method is based on the classical view of turbulence as a superposition of coherent structures. It is able to reproduce prescribed first and second order one point statistics, characteristic length and time scales, and the shape of coherent structures. The ability of the method to produce realistic inflow conditions in the test cases of a spatially decaying homogeneous isotropic turbulence and of a fully developed turbulent channel flow is presented. The method is systematically compared to other methods of generation of inflow conditions (precursor simulation, spectral methods and algebraic methods). © 2006 Elsevier Inc. All rights reserved.},
+   author = {N. Jarrin and S. Benhamadouche and D. Laurence and R. Prosser},
+   doi = {10.1016/j.ijheatfluidflow.2006.02.006},
+   issn = {0142727X},
+   issue = {4},
+   journal = {International Journal of Heat and Fluid Flow},
+   keywords = {Channel flow,Inflow boundary conditions,Large-eddy simulation,Spatially decaying isotropic turbulence,Synthetic turbulence},
+   month = {8},
+   pages = {585-593},
+   title = {A synthetic-eddy-method for generating inflow conditions for large-eddy simulations},
+   volume = {27},
+   year = {2006},
+}
+
+@article{Poletto2013,
+   author = {R. Poletto and T. Craft and A. Revell},
+   doi = {10.1007/s10494-013-9488-2},
+   issn = {13866184},
+   issue = {3},
+   journal = {Flow, Turbulence and Combustion},
+   keywords = {Embedded LES,SEM,Synthetic Eddy method,Synthetic turbulence},
+   month = {10},
+   pages = {519-539},
+   title = {A new divergence free synthetic eddy method for the reproduction of inlet flow conditions for les},
+   volume = {91},
+   year = {2013},
+}
+
+@article{OH2019,
+title = {Extended synthetic eddy method to generate inflow data for turbulent thermal boundary layer},
+journal = {International Journal of Heat and Mass Transfer},
+volume = {134},
+pages = {1261-1267},
+year = {2019},
+issn = {0017-9310},
+doi = {10.1016/j.ijheatmasstransfer.2019.02.061},
+url = {https://www.sciencedirect.com/science/article/pii/S0017931018332289},
+author = {Geunwoo Oh and Kyung Min Noh and Hyunwook Park and Jung-Il Choi},
+keywords = {Synthetic eddy method, Inflow boundary condition, Turbulent thermal boundary layer, Large-eddy simulation},
+}
