+++ README.md
@@ -1,6 +1,6 @@
 
 # Synthetic Eddy Method 
-<img src="https://github.com/carlodev/SyntheticEddyMethod.jl/blob/master/images/logo/logo.png" width="100" title="SEM logo">
+<img src="https://github.com/carlodev/SyntheticEddyMethod.jl/blob/master/paper/images/logo/logo.png" width="100" title="SEM logo">
 
 **Documentation**
 
