### 2023-07-20 #5565 @carlodev
Hi @philipcardiff, I have merged your pull request, fixed the rendering error pointed out by @akshaysridhar .
I merged the joss folder with the master, published a new version (0.4.2) and it has been archived on zenodo: 10.5281/zenodo.8167205
