### 2023-07-20 8857b02
Update paper.md: fix minor typo
