### 2023-07-20 #5565 @philipcardiff
Hi @carlodev , I made one more typo fix at https://github.com/carlodev/SyntheticEddyMethod.jl/pull/22/files.
