### 2023-07-20 #5565 @carlodev
I have checked it, and changed a capital letter in a figure caption
